+++ packages/syntax-core/src/AvatarGroup/AvatarGroup.test.tsx
@@ -1,15 +1,12 @@
 import { screen, render } from "@testing-library/react";
 import AvatarGroup from "./AvatarGroup";
+import Avatar from "../Avatar/Avatar";
 
 describe("avatarGroup", () => {
   it("renders successfully", () => {
     const baseElement = render(
       <AvatarGroup>
-        <AvatarGroup.Avatar
-          accessibilityLabel="Joseph Liotta"
-          src="image.png"
-          zIndex={3}
-        />
+        <Avatar accessibilityLabel="Joseph Liotta" src="image.png" />
       </AvatarGroup>,
     );
     expect(baseElement).toBeTruthy();
@@ -17,21 +14,9 @@ describe("avatarGroup", () => {
   it("renders images with the correct labels", async () => {
     render(
       <AvatarGroup>
-        <AvatarGroup.Avatar
-          accessibilityLabel="Joseph Liotta"
-          src="image.png"
-          zIndex={3}
-        />
-        <AvatarGroup.Avatar
-          accessibilityLabel="Empty Avatar 1"
-          src="image.png"
-          zIndex={2}
-        />
-        <AvatarGroup.Avatar
-          accessibilityLabel="Empty Avatar 2"
-          src="image.png"
-          zIndex={1}
-        />
+        <Avatar accessibilityLabel="Joseph Liotta" src="image.png" />
+        <Avatar accessibilityLabel="Empty Avatar 1" src="image.png" />
+        <Avatar accessibilityLabel="Empty Avatar 2" src="image.png" />
       </AvatarGroup>,
     );
     const image1 = await screen.findByAltText("Joseph Liotta");